refactor(provider): move SES helpers out of provider.go; schema: fix dynamo descriptions; add AuthTable* outputs and deprecate TenantTable*; alphabetize outputs; emit new outputs in provider
